Change distance type to double

+++ src/org/usfirst/frc/team3090/robot/Distance.java
@@ -2,9 +2,9 @@ package org.usfirst.frc.team3090.robot;
 
 public class Distance extends Step {
 
-	public final float inches;
+	public final double inches;
 
-	public Distance(float inches) {
+	public Distance(double inches) {
 		this.inches = inches;
 	}
